starting page, solo file needed
On this site you type in a message, give the message a name, and give the message a key. Another user can then visit the site who has knowledge of the file name and key, and view the message if they type in that information. A txt file is created on the server.

The file created that contains their message is an AES encrypted txt file on the server.